+++ gui/src/main/kotlin/dev/robocode/tankroyale/gui/ui/arena/ArenaPanel.kt
@@ -122,8 +122,10 @@ object ArenaPanel : JPanel() {
             TankColorMode.BOT_COLORS -> fromString(botColor ?: default)
             TankColorMode.BOT_COLORS_ONCE -> {
                 val botLockedColors = lockedColors.getOrPut(ownerId) { mutableMapOf() }
-                val lockedColor = botLockedColors.getOrPut(default) { botColor ?: "" }
-                fromString(if (lockedColor.isEmpty()) default else lockedColor)
+                if (!botLockedColors.containsKey(default) && botColor != null) {
+                    botLockedColors[default] = botColor
+                }
+                fromString(botLockedColors[default] ?: botColor ?: default)
             }
             TankColorMode.DEFAULT_COLORS -> fromString(default)
             TankColorMode.BOT_COLORS_WHEN_DEBUGGING ->